refactor(compat): Moved backward-compatibility functionality to it's own module
Minor performance improvements.
